move alerts away from state management
