+++ libs/client/alerts/src/lib/alerts.module.ts
@@ -1,19 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { NgxsModule } from '@ngxs/store';
 import { IconsModule } from '@dragonfish/client/icons';
-import { AlertsService, AlertsState } from './shared';
-import { AlertsComponent } from './alerts.component'
+import { AlertsService } from './alerts.service';
+import { AlertsComponent } from './alerts.component';
 
 @NgModule({
     declarations: [AlertsComponent],
-    imports: [
-        CommonModule,
-        IconsModule,
-        NgxsModule.forFeature([AlertsState]),
-    ],
-    providers: [AlertsService]
+    imports: [CommonModule, IconsModule],
+    providers: [AlertsService],
 })
 export class AlertsModule {}
 
-export { AlertsService } from './shared';
+export { AlertsService } from './alerts.service';
