--- libs/client/alerts/src/lib/shared/index.ts
@@ -1,4 +1,0 @@
-export * from './services';
-
-export * from './alerts.actions';
-export { AlertsState } from './alerts.state';
