--- libs/client/alerts/src/lib/shared/services/index.ts
@@ -1,1 +1,0 @@
-export { AlertsService } from './alerts.service';
